+++ hutool-core/src/main/java/cn/hutool/core/date/ChineseDate.java
@@ -21,7 +21,7 @@ public class ChineseDate {
 	/**
 	 * 1900-01-31
 	 */
-	private static final long baseDate = -2206425943000L;
+	private static final long BASE_DATE = -2206425943000L;
 	//农历年
 	private final int year;
 	//农历月
@@ -46,7 +46,7 @@ public class ChineseDate {
 	 */
 	public ChineseDate(Date date) {
 		// 求出和1900年1月31日相差的天数
-		int offset = (int) ((date.getTime() - baseDate) / DateUnit.DAY.getMillis());
+		int offset = (int) ((date.getTime() - BASE_DATE) / DateUnit.DAY.getMillis());
 		// 计算农历年份
 		// 用offset减去每农历年的天数，计算当天是农历第几天，offset是当年的第几天
 		int daysOfYear;
@@ -299,7 +299,7 @@ public class ChineseDate {
 		if (D >= firstNode) {
 			gzM = GanZhi.cyclicalm((Y - LunarInfo.BASE_YEAR) * 12 + M + 12);
 		}
-		int dayCyclical = (int) ((DateUtil.parseDate(Y + "-" + M + "-" + "1").getTime() - baseDate + 2592000000L) / DateUnit.DAY.getMillis()) + 10;
+		int dayCyclical = (int) ((DateUtil.parseDate(Y + "-" + M + "-" + "1").getTime() - BASE_DATE + 2592000000L) / DateUnit.DAY.getMillis()) + 10;
 		String gzD = GanZhi.cyclicalm(dayCyclical + D - 1);
 		return gzyear + "年" + gzM + "月" + gzD + "日";
 	}