+++ hutool-core/src/main/java/cn/hutool/core/util/CharUtil.java
@@ -11,24 +11,43 @@ import cn.hutool.core.text.ASCIIStrCache;
  */
 public class CharUtil {
 
+	/** 字符常量：空格符 ' ' */
 	public static final char SPACE = ' ';
+	/** 字符常量：制表符 \t */
 	public static final char TAB = '	';
+	/** 字符常量：点 . */
 	public static final char DOT = '.';
+	/** 字符常量：斜杠 / */
 	public static final char SLASH = '/';
+	/** 字符常量：反斜杠 \ */
 	public static final char BACKSLASH = '\\';
+	/** 字符常量：回车符 \r */
 	public static final char CR = '\r';
+	/** 字符常量：换行符 \n */
 	public static final char LF = '\n';
-	public static final char UNDERLINE = '_';
+	/** 字符常量：连接符 - */
 	public static final char DASHED = '-';
+	/** 字符常量：下划线 _ */
+	public static final char UNDERLINE = '_';
+	/** 字符常量：逗号 , */
 	public static final char COMMA = ',';
+	/** 字符常量：花括号（左） { */
 	public static final char DELIM_START = '{';
+	/** 字符常量：花括号（右） } */
 	public static final char DELIM_END = '}';
+	/** 字符常量：中括号（左） [ */
 	public static final char BRACKET_START = '[';
+	/** 字符常量：中括号（右） ] */
 	public static final char BRACKET_END = ']';
-	public static final char COLON = ':';
+	/** 字符常量：双引号 : */
 	public static final char DOUBLE_QUOTES = '"';
+	/** 字符常量：单引号 ' */
 	public static final char SINGLE_QUOTE = '\'';
+	/** 字符常量：与 &amp; */
 	public static final char AMP = '&';
+	/** 字符常量：冒号 : */
+	public static final char COLON = ':';
+	/** 字符常量：艾特 @ */
 	public static final char AT = '@';
 
 	/**