+++ hutool-core/src/test/java/cn/hutool/core/util/NumberUtilTest.java
@@ -241,15 +241,32 @@ public class NumberUtilTest {
 		long factorial = NumberUtil.factorial(0);
 		Assert.assertEquals(1, factorial);
 
+		Assert.assertEquals(1L, NumberUtil.factorial(1));
+		Assert.assertEquals(1307674368000L, NumberUtil.factorial(15));
+		Assert.assertEquals(2432902008176640000L, NumberUtil.factorial(20));
+
 		factorial = NumberUtil.factorial(5, 0);
 		Assert.assertEquals(120, factorial);
 		factorial = NumberUtil.factorial(5, 1);
 		Assert.assertEquals(120, factorial);
+    
+		Assert.assertEquals(5, NumberUtil.factorial(5, 4));
+		Assert.assertEquals(2432902008176640000L, NumberUtil.factorial(20, 0));
 	}
 
 	@Test
 	public void mulTest(){
 		final BigDecimal mul = NumberUtil.mul(new BigDecimal("10"), null);
 		Assert.assertEquals(BigDecimal.ZERO, mul);
+	}
+	
+	
+	@Test
+	public void isPowerOfTwoTest() {
+		Assert.assertEquals(false, NumberUtil.isPowerOfTwo(-1));
+		Assert.assertEquals(true, NumberUtil.isPowerOfTwo(16));
+		Assert.assertEquals(true, NumberUtil.isPowerOfTwo(65536));
+		Assert.assertEquals(true, NumberUtil.isPowerOfTwo(1));
+		Assert.assertEquals(false, NumberUtil.isPowerOfTwo(17));
 	}
 }