+++ hutool-dfa/src/main/java/cn/hutool/dfa/SensitiveUtil.java
@@ -1,5 +1,6 @@
 package cn.hutool.dfa;
 
+import cn.hutool.core.lang.Filter;
 import cn.hutool.core.thread.ThreadUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.json.JSONUtil;
@@ -13,8 +14,7 @@ import java.util.List;
  *
  */
 public final class SensitiveUtil {
-//	private static final Log log = LogFactory.get();
-	
+
 	public static final char DEFAULT_SEPARATOR = StrUtil.C_COMMA;
 	private static final WordTree sensitiveTree = new WordTree();
 	
@@ -70,6 +70,19 @@ public final class SensitiveUtil {
 	 */
 	public static void init(String sensitiveWords, boolean isAsync){
 		init(sensitiveWords, DEFAULT_SEPARATOR, isAsync);
+	}
+	
+	/**
+	 * 设置字符过滤规则，通过定义字符串过滤规则，过滤不需要的字符<br>
+	 * 当accept为false时，此字符不参与匹配
+	 *
+	 * @param charFilter 过滤函数
+	 * @since 5.4.4
+	 */
+	public static void setCharFilter(Filter<Character> charFilter) {
+		if(charFilter != null) {
+			sensitiveTree.setCharFilter(charFilter);
+		}
 	}
 	
 	/**